Fix warning.

+++ src/sysinfo/nvapi_adapter_registry.cpp
@@ -82,7 +82,7 @@ namespace dxvk {
     short NvapiAdapterRegistry::GetPrimaryOutputId() const {
         for (auto i = 0U; i < m_nvapiOutputs.size(); i++)
             if (m_nvapiOutputs.at(i)->IsPrimary())
-                return i;
+                return static_cast<short>(i);
 
         return -1;
     }
@@ -90,7 +90,7 @@ namespace dxvk {
     short NvapiAdapterRegistry::GetOutputId(const std::string& displayName) const {
         for (auto i = 0U; i < m_nvapiOutputs.size(); i++)
             if (m_nvapiOutputs.at(i)->GetDeviceName() == displayName)
-                return i;
+                return static_cast<short>(i);
 
         return -1;
     }
